# backend/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateBidRequest:
    properties:
      answers:
        additionalProperties:
          type: string
        type: object
      estimated_completion:
        type: string
      message:
        minLength: 10
        type: string
    required:
    - estimated_completion
    - message
    type: object
  models.CreateTaskRequest:
    properties:
      deadline:
        type: string
      description:
        minLength: 10
        type: string
      org_id:
        type: string
      priority:
        allOf:
        - $ref: '#/definitions/models.TaskPriority'
        enum:
        - low
        - medium
        - high
        - critical
      questions:
        items:
          type: string
        type: array
      skills:
        items:
          type: string
        type: array
      title:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - deadline
    - description
    - priority
    - title
    type: object
...
  models.RateTaskRequest:
    properties:
      points:
        maximum: 10000
        minimum: 0
        type: integer
      rating:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - points
    - rating
    type: object
...
  models.TaskPriority:
    enum:
    - low
    - medium
    - high
    - critical
    type: string
    x-enum-varnames:
    - PriorityLow
    - PriorityMedium
    - PriorityHigh
    - PriorityCritical
  models.TaskStatus:
    enum:
    - open
    - assigned
    - in_progress
    - completed
    - closed
    type: string
    x-enum-varnames:
    - StatusOpen
    - StatusAssigned
    - StatusInProgress
    - StatusCompleted
    - StatusClosed
...
  models.UpdateTaskRequest:
    properties:
      deadline:
        type: string
      description:
        minLength: 10
        type: string
      priority:
        allOf:
        - $ref: '#/definitions/models.TaskPriority'
        enum:
        - low
        - medium
        - high
        - critical
      questions:
        items:
          type: string
        type: array
      skills:
        items:
          type: string
        minItems: 1
        type: array
      status:
        allOf:
        - $ref: '#/definitions/models.TaskStatus'
        enum:
        - open
        - assigned
        - in_progress
        - completed
        - closed
      title:
        maxLength: 255
        minLength: 3
        type: string
    type: object
info:
  contact: {}
  description: API documentation for Task Delegation Platform.
  title: Task Delegation Platform API
  version: "1.0"
paths:
...
  /bids/{id}/approve:
    patch:
      parameters:
      - description: Bid ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Approve a bid
      tags:
      - bids
  /bids/{id}/reject:
    patch:
      parameters:
      - description: Bid ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Reject a bid
      tags:
      - bids
  /bids/my:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List current user's bids
      tags:
      - bids
...
  /tasks:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Task status
        in: query
        name: status
        type: string
      - description: Task priority
        in: query
        name: priority
        type: string
      - description: Sort key
        in: query
        name: sort
        type: string
      - collectionFormat: csv
        description: Skills filter
        in: query
        items:
          type: string
        name: skills
        type: array
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Search and list tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: Create task request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete task
      tags:
      - tasks
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get task detail
      tags:
      - tasks
    put:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Update task request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update task
      tags:
      - tasks
  /tasks/{id}/bids:
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List bids for task
      tags:
      - bids
    post:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Create bid request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateBidRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Place bid on task
      tags:
      - bids
  /tasks/{id}/checklist:
    put:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Checklist update request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update task checklist
      tags:
      - tasks
  /tasks/{id}/comments:
    post:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment body request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add task comment
      tags:
      - tasks
  /tasks/{id}/rate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Rate task request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Rate completed task
      tags:
      - tasks
  /tasks/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Status transition request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Transition task status
      tags:
      - tasks
  /tasks/my:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user's tasks
      tags:
      - tasks
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"